backend/requirements.txt
# 核心依赖
flask
flask-cors
python-dotenv
psycopg2-binary

# 文档处理
python-multipart
markdown

# AI 集成
openai
anthropic

# 部署
gunicorn

# 测试
pytest
pytest-cov